requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic-settings==2.5.0
